Ruffification

+++ wcivf/apps/core/management/commands/init_data.py
@@ -2,13 +2,12 @@ import os
 
 from django.conf import settings
 from django.core.cache import cache
-from django.core.management.base import BaseCommand
 from django.core.management import call_command
-
+from django.core.management.base import BaseCommand
 from elections.constants import (
     PEOPLE_FOR_BALLOT_KEY_FMT,
-    POSTCODE_TO_BALLOT_KEY_FMT,
     POLLING_STATIONS_KEY_FMT,
+    POSTCODE_TO_BALLOT_KEY_FMT,
 )
 
 
@@ -36,15 +35,14 @@ class Command(BaseCommand):
             print(" ".join(command))
             call_command(*command)
 
-        if options["full"]:
-            # Delete the cache on a full import
-            if hasattr(cache, "delete_pattern"):
-                for fmt in (
-                    POLLING_STATIONS_KEY_FMT,
-                    POSTCODE_TO_BALLOT_KEY_FMT,
-                    PEOPLE_FOR_BALLOT_KEY_FMT,
-                ):
-                    cache.delete_pattern(fmt.format("*"))
+        # Delete the cache on a full import
+        if options["full"] and hasattr(cache, "delete_pattern"):
+            for fmt in (
+                POLLING_STATIONS_KEY_FMT,
+                POSTCODE_TO_BALLOT_KEY_FMT,
+                PEOPLE_FOR_BALLOT_KEY_FMT,
+            ):
+                cache.delete_pattern(fmt.format("*"))
 
         # Unset dirty file if it exists
         if getattr(settings, "CHECK_HOST_DIRTY", False):
